Remove index.js of routes to follow the routes implementation at other apps
