+++ src/validation.js
@@ -26,7 +26,7 @@ import { getRequestToken } from '@nextcloud/auth'
 import Vue from 'vue'
 
 import Validation from './views/Validation.vue'
-import router from './router/index.js'
+import router from './router/router.js'
 import store from './store/index.js'
 
 // CSP config for webpack dynamic chunk loading
@@ -45,12 +45,6 @@ Vue.prototype.n = n
 
 Vue.prototype.OC = OC
 Vue.prototype.OCA = OCA
-
-if (
-	window.location.pathname.split('/')[1] === 'index.php' && OC.config.modRewriteWorking
-) {
-	router.push({ name: 'Validation' })
-}
 
 export default new Vue({
 	el: '#content',
